+++ src/compas_viewer/components/dialog.py
@@ -48,13 +48,13 @@ class CameraSettingsDialog(QDialog, Base):
         super().__init__()
         self.setWindowTitle("Camera Settings")
 
-        # Create layout for spin boxes
         self.layout = QVBoxLayout(self)
         self.spin_boxes = {}
         current_camera = self.viewer.renderer.camera
+        # set None to infinity error
         coordinates = {
-            "Camera_Target": [("X", current_camera.target.x, -10000, 10000), ("Y", current_camera.target.y, -10000, 10000), ("Z", current_camera.target.z, -10000, 10000)],
-            "Camera_Position": [("X", current_camera.position.x, -10000, 10000), ("Y", current_camera.position.y, -10000, 10000), ("Z", current_camera.position.z, -10000, 10000)],
+            "Camera_Target": [("X", current_camera.target.x, None, None), ("Y", current_camera.target.y, None, None), ("Z", current_camera.target.z, None, None)],
+            "Camera_Position": [("X", current_camera.position.x, None, None), ("Y", current_camera.position.y, None, None), ("Z", current_camera.position.z, None, None)],
         }
 
         for coord in coordinates: