+++ src/compas_viewer/components/double_edit.py
@@ -1,9 +1,17 @@
+import sys
+
 from PySide6 import QtWidgets
 
 
 class DoubleEdit(QtWidgets.QWidget):
-    def __init__(self, name: str, default: float, min_val: float, max_val: float):
+    def __init__(self, name: str = None, default: float = None, min_val: float = None, max_val: float = None):
         super().__init__()
+
+        if min_val is None:
+            min_val = -sys.float_info.max
+        if max_val is None:
+            max_val = sys.float_info.max
+
         self.layout = QtWidgets.QHBoxLayout()
         self.label = QtWidgets.QLabel(name)
         self.spinbox = QtWidgets.QDoubleSpinBox()