+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fixed `action` bug.
 * Changed `GridObject` to draw direcly in `Render`
-* Fixed `CameraSettingsDialog` updat current position bug.
+* Fixed `CameraSettingsDialog` update current position bug.
 
 ### Removed